% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codons_to_aas.R
\name{codons_to_aas}
\alias{codons_to_aas}
\title{Codons to amino acid sequence}
\usage{
codons_to_aas(codons)
}
\arguments{
\item{codons}{list/vector of codons (as strings)}
}
\value{
an amino acid sequence resulting from the provided codons.
}
\description{
Codons to amino acid sequence
}
\examples{
codons_to_aas(c("UUU", "UCU"))

}
