% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aa_composition.R
\name{plot_aa_frequency}
\alias{plot_aa_frequency}
\title{Plot frequency of amino acids}
\usage{
plot_aa_frequency(sequence)
}
\arguments{
\item{sequence}{A string representing an amino acid sequence}
}
\value{
A plot showing the frequency of each amino acid in the sequence
}
\description{
Plot frequency of amino acids
}
\section{Functions}{
\itemize{
\item \code{plot_aa_frequency()}: a bar plot showing the frequency of each amino acid in a protein sequence.

}}
\examples{
# Example 1: Simple peptide sequence
plot_aa_frequency("ACDEFGHIKLMNPQRSTVWY")

# Example 2: Repeated amino acids
plot_aa_frequency("AAAAGGTTCC")
}
