% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_to_rna.R
\name{dna_to_rna}
\alias{dna_to_rna}
\title{Convert DNA sequence to RNA sequence}
\usage{
dna_to_rna(dna_seq)
}
\arguments{
\item{dna_seq}{A character string representing a DNA sequence.}
}
\value{
A character string containing the corresponding RNA sequence.
}
\description{
This function replaces all occurrences of thymine ("T") with uracil ("U"),
converting a DNA sequence into its corresponding RNA sequence.
}
\examples{
dna_to_rna("ATGCTTACG")
dna_to_rna("TTTAAA")
}
