% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_codons.R
\name{extract_codons}
\alias{extract_codons}
\title{Split a nucleotide sequence into codons}
\usage{
extract_codons(sequence, start = 1)
}
\arguments{
\item{sequence}{A character string representing a nucleotide sequence.}

\item{start}{An integer indicating the starting position for reading codons.}
}
\value{
A character vector where each element is a codon (a triplet of nucleotides) extracted from the input sequence.
}
\description{
Split a nucleotide sequence into codons
}
\examples{
# Split a DNA sequence into codons
extract_codons("ATGCGATAA")

# Start reading from the second position
extract_codons("ATGCGATAA", start = 2)
}
