% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dna_sequence.R
\name{generate_dna_sequence}
\alias{generate_dna_sequence}
\title{Generate a random DNA sequence}
\usage{
generate_dna_sequence(length)
}
\arguments{
\item{length}{Integer. The number of nucleotide bases in the generated DNA sequence.}
}
\value{
A character string representing the random DNA sequence of the specified length.
}
\description{
This function generates a random DNA sequence composed of the nucleotides
adenine (A), thymine (T), guanine (G), and cytosine (C). The user specifies
the desired sequence length.
}
\examples{
# Generate a random DNA sequence of length 10
generate_dna_sequence(10)

# Generate a longer sequence of length 100
generate_dna_sequence(100)
}
